backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.3.2
PyYAML==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
gitpython==3.1.41
python-dotenv==1.0.0
bandit==1.7.6
safety==3.2.0
semgrep==1.55.0
